django/requirements.txt
boto3
botocore
Django
django-storages
celery
future
mysqlclient
opencv-python
requests
PythonVideoConverter
gunicorn
